requirements.txt
certifi==2018.8.24
cfe==0.0.15
Django==2.1.1
django-markdown-deux==1.0.5
django-pagedown==1.0.6
gunicorn==19.9.0
markdown2==2.3.6
Pillow==5.3.0
pipenv==2018.7.1
psycopg2-binary==2.7.5
pytz==2018.5
virtualenv==16.0.0
virtualenv-clone==0.3.0
